[CST-9636] Show warning only when there are no controls and form is not disabled
