+++ src/app/shared/access-control-form-container/access-control-array-form/access-control-array-form.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, NgModule, OnDestroy, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormArray, FormBuilder, FormControl, ReactiveFormsModule } from '@angular/forms';
-import { SharedBrowseByModule } from '../browse-by/shared-browse-by.module';
+import { SharedBrowseByModule } from '../../browse-by/shared-browse-by.module';
 import { TranslateModule } from '@ngx-translate/core';
 import { NgbDatepickerModule } from '@ng-bootstrap/ng-bootstrap';
 import { ControlMaxStartDatePipe } from './control-max-start-date.pipe';
@@ -9,7 +9,7 @@ import { ControlMaxEndDatePipe } from './control-max-end-date.pipe';
 
 import { distinctUntilChanged, takeUntil } from 'rxjs/operators';
 import { Subject } from 'rxjs';
-import { AccessesConditionOption } from '../../core/config/models/config-accesses-conditions-options.model';
+import { AccessesConditionOption } from '../../../core/config/models/config-accesses-conditions-options.model';
 
 
 @Component({
