+++ src/app/shared/access-control-form-container/access-control-array-form/control-max-start-date.pipe.ts
@@ -1,21 +1,22 @@
 import { Pipe, PipeTransform } from '@angular/core';
 import { AbstractControl } from '@angular/forms';
 import { NgbDateStruct } from '@ng-bootstrap/ng-bootstrap/datepicker/ngb-date-struct';
-import { AccessesConditionOption } from '../../core/config/models/config-accesses-conditions-options.model';
+import { AccessesConditionOption } from '../../../core/config/models/config-accesses-conditions-options.model';
+
 
 @Pipe({
   // eslint-disable-next-line @angular-eslint/pipe-prefix
-  name: 'maxEndDate',
+  name: 'maxStartDate',
   pure: false
 })
-export class ControlMaxEndDatePipe implements PipeTransform {
+export class ControlMaxStartDatePipe implements PipeTransform {
   transform(control: AbstractControl, dropdownOptions: AccessesConditionOption[]): NgbDateStruct | null {
     const { itemName } = control.value;
     const item = dropdownOptions.find((x) => x.name === itemName);
-    if (!item?.hasEndDate) {
+    if (!item?.hasStartDate) {
       return null;
     }
-    const date = new Date(item.maxEndDate);
+    const date = new Date(item.maxStartDate);
     return {
       year: date.getFullYear(),
       month: date.getMonth() + 1,
