+++ src/app/shared/access-control-form-container/access-control-form-container.component.html
@@ -41,12 +41,7 @@
             </div>
           </div>
 
-          <div class="mt-3">
-            <h5>{{'access-control-access-conditions' | translate}}</h5>
-            <div class="alert alert-warning">
-              {{'access-control-no-access-conditions-warning-message' | translate}}
-            </div>
-          </div>
+          <h5 class="mt-3">{{'access-control-access-conditions' | translate}}</h5>
 
           <ds-access-control-array-form
             #itemAccessCmp
@@ -127,13 +122,7 @@
             </div>
           </div>
 
-          <div class="mt-3">
-            <h5>{{'access-control-access-conditions' | translate}}</h5>
-
-            <div class="alert alert-warning">
-              {{'access-control-no-access-conditions-warning-message' | translate}}
-            </div>
-          </div>
+          <h5 class="mt-3">{{'access-control-access-conditions' | translate}}</h5>
 
           <ds-access-control-array-form
             #bitstreamAccessCmp
